restructure stubs and testbeds classes in auth feature tests
